requirements.txt
asttokens==2.4.1
attrs==23.2.0
black==24.3.0
cfgv==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
duckdb==0.10.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flake8==7.0.0
identify==2.5.35
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
nbformat==5.10.3
nbstripout==0.7.1
nest-asyncio==1.6.0
nodeenv==1.8.0
packaging==24.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
# Editable Git install with no remote (src==0.0.1)
-e /home/hamedbh/projects/multilingual-example
stack-data==0.6.3
tokenize-rt==5.2.0
tomli==2.0.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
virtualenv==20.25.1
wcwidth==0.2.13
